+++ bioinformatics-software-framework/src/app/components/step-box/step-box.component.css
